// testcharmvision-ts/client/src/api/types/BrowserSessionResponse.ts
// This file was auto-generated by Fern from our API Definition.

import type * as Testcharmvision from "../index.js";

/**
 * Response model for browser session information.
 */
export interface BrowserSessionResponse {
    /** Unique identifier for the browser session. browser_session_id starts with `pbs_`. */
    browser_session_id: string;
    /** ID of the organization that owns this session */
    organization_id: string;
    /** Current status of the browser session */
    status?: string;
    /** Type of the current runnable associated with this session (workflow, task etc) */
    runnable_type?: string;
    /** ID of the current runnable */
    runnable_id?: string;
    /** Timeout in minutes for the session. Timeout is applied after the session is started. Defaults to 60 minutes. */
    timeout?: number;
    /** Url for connecting to the browser */
    browser_address?: string;
    /** Url for the browser session page */
    app_url?: string;
    /** A list of extensions installed in the browser session. */
    extensions?: Testcharmvision.Extensions[];
    /** The type of browser used for the session. */
    browser_type?: Testcharmvision.PersistentBrowserType;
    /** Whether the browser session supports VNC streaming */
    vnc_streaming_supported?: boolean;
    /** The path where the browser session downloads files */
    download_path?: string;
    /** The list of files downloaded by the browser session */
    downloaded_files?: Testcharmvision.FileInfo[];
    /** The list of video recordings from the browser session */
    recordings?: Testcharmvision.FileInfo[];
    /** Timestamp when the session was started */
    started_at?: string;
    /** Timestamp when the session was completed */
    completed_at?: string;
    /** Timestamp when the session was created (the timestamp for the initial request) */
    created_at: string;
    /** Timestamp when the session was last modified */
    modified_at: string;
    /** Timestamp when the session was deleted, if applicable */
    deleted_at?: string;
}