// testcharmvision-ts/client/src/environments.ts
// This file was auto-generated by Fern from our API Definition.

export const TestcharmvisionEnvironment = {
    Cloud: "https://api.testcharmvision.com",
    Staging: "https://api-staging.testcharmvision.com",
    Local: "http://localhost:8000",
} as const;

export type TestcharmvisionEnvironment =
    | typeof TestcharmvisionEnvironment.Cloud
    | typeof TestcharmvisionEnvironment.Staging
    | typeof TestcharmvisionEnvironment.Local;